fix: pr feedback - co-locating types
