+++ src/components/Common/Fields/RadioGroupField/RadioGroupField.tsx
@@ -1,6 +1,6 @@
 import { useField, type UseFieldProps } from '@/components/Common/Fields/hooks/useField'
-import { RadioGroup, type RadioGroupProps } from '@/components/Common/UI/RadioGroup'
-
+import { RadioGroup } from '@/components/Common/UI/RadioGroup'
+import type { RadioGroupProps } from '@/components/Common/UI/RadioGroup/RadioGroupTypes'
 export interface RadioGroupFieldProps
   extends Omit<RadioGroupProps, 'value'>,
     UseFieldProps<string> {}
