+++ src/components/Common/UI/CheckboxGroup/index.ts
@@ -1,1 +1,1 @@
-export { CheckboxGroup, type CheckboxGroupProps, type CheckboxGroupOptions } from './CheckboxGroup'
+export { CheckboxGroup } from './CheckboxGroup'
