+++ src/components/Common/UI/HorizontalFieldLayout/HorizontalFieldLayout.tsx
@@ -3,8 +3,8 @@ import { VisuallyHidden } from 'react-aria'
 import { FieldErrorMessage } from '../FieldErrorMessage'
 import { FieldDescription } from '../FieldDescription'
 import styles from './HorizontalFieldLayout.module.scss'
+import type { HorizontalFieldLayoutProps } from './HorizontalFieldLayoutTypes'
 import { getDataProps } from '@/helpers/getDataProps'
-import type { HorizontalFieldLayoutProps } from '@/types/UI/HorizontalFieldLayout'
 
 export const HorizontalFieldLayout: React.FC<HorizontalFieldLayoutProps> = ({
   label,
