+++ src/components/Common/UI/RadioGroup/RadioGroupTypes.ts
@@ -1,0 +1,20 @@
+import type { FieldsetHTMLAttributes, Ref } from 'react'
+import type { SharedFieldLayoutProps } from '../FieldLayout/FieldLayoutTypes'
+
+export type RadioGroupOptions = {
+  label: React.ReactNode
+  value: string
+  isDisabled?: boolean
+  description?: React.ReactNode
+}
+
+export interface RadioGroupProps
+  extends SharedFieldLayoutProps,
+    Pick<FieldsetHTMLAttributes<HTMLFieldSetElement>, 'className'> {
+  isInvalid?: boolean
+  isDisabled?: boolean
+  options: Array<RadioGroupOptions>
+  value?: string
+  onChange?: (value: string) => void
+  inputRef?: Ref<HTMLInputElement>
+}
