+++ src/components/Common/UI/RadioGroup/index.ts
@@ -1,2 +1,1 @@
 export { RadioGroup } from './RadioGroup'
-export type { RadioGroupProps, RadioGroupOptions } from './RadioGroup'
